Bumped Java Client to 6.4.1
Got test working for only metadata-values and removed now-unnecessary cast for GenericDocumentManager.

+++ src/test/java/com/marklogic/mule/extension/MetadataVerifier.java
@@ -135,7 +135,7 @@ public class MetadataVerifier {
 
     private void verifyMetadataValues() {
         if (expectedMetadataValueCount != null) {
-            assertEquals(((long) expectedMetadataValueCount), attributes.getMetadataValues().size());
+            assertEquals(((int) expectedMetadataValueCount), attributes.getMetadataValues().size());
         }
         if (expectedMetadataValues != null) {
             for (int i = 0; i < expectedMetadataValues.length; i = i + 2) {
